created /lib and migrated overwrite-random.js
